Refatora estrutura de schemas e especializa endpoints, conforme issue #68
- Remove schema genérico PreencherSolicitacaoInput e adota schemas especializados (RequisicaoMamografiaRastreamentoSchema e RequisicaoMamografiaDiagnosticaSchema) nos respectivos endpoints
- Centraliza LoginInput em src/siscan/schema/auth_schema.py e ajusta endpoints de autenticação
- Adiciona exemplos completos (json_schema_extra) nos schemas especializados, aprimorando documentação da API
- Cria src/siscan/webpage/types.py para InputType, atualizando módulos de automação web
- Remove campos duplicados em RequisicaoNovoExameSchema e ajusta validações
- Atualiza documentação interna e comentários explicativos

Refs #68
