+++ src/siscan/requisicao/requisicao_exame_mamografia_rastreio.py
@@ -1,3 +1,10 @@
+#
+# Subclasse de RequisicaoExameMamografia, focada na lógica para mamografia de
+# rastreamento.
+#
+# Define campos e métodos específicos desse subtipo de requisição
+# (preenchimento de tipo, rastreio, responsável pela coleta etc.)
+#
 import logging
 from pydantic import BaseModel
 from typing import Type, Any
