+++ src/utils/validator.py
@@ -1,13 +1,10 @@
 from __future__ import annotations
 
-import logging
-
 import json
+import logging
 from pathlib import Path
-from typing import Any, Dict, List, Union, Type
-
 from pydantic import BaseModel, ValidationError
-
+from typing import Any, Dict, List, Union, Type
 
 logger = logging.getLogger(__name__)
